cosmosdb/crud_app/requirements.txt
SQLAlchemy
python-dotenv
psycopg2
azure-identity
alembic